validaciones en el menu de pacientes y proveedores

+++ Clinica_Medica_Polanco/AgregarEmpleado.xaml.cs
@@ -56,10 +56,10 @@ namespace Clinica_Medica_Polanco
                 empleados1.TelefonoEmpleado = txt_Telefono.Text;
                 empleados1.FechaNacimientoEmpleado = Convert.ToDateTime(dtp_Nacimiento.Text);
                 empleados1.CorreoEmpleado = txt_Correo.Text;
-                empleados1.AlturaEmpleado = string.IsNullOrEmpty(txt_Altura.Text) ? 0 : int.Parse(txt_Altura.Text);
+                empleados1.AlturaEmpleado = string.IsNullOrEmpty(txt_Altura.Text) ? 0 : decimal.Parse(txt_Altura.Text);
                 empleados1.TipoSangreEmpleado = cmb_Tipo_Sangre.SelectedItem.ToString();
                 empleados1.DireccionEmpleado = string.IsNullOrWhiteSpace(rtbAString(rtb_Direccion)) ? null : rtbAString(rtb_Direccion);
-                empleados1.SueldoBase = string.IsNullOrEmpty(txt_Sueldo.Text) ? -1 : int.Parse(txt_Sueldo.Text);
+                empleados1.SueldoBase = string.IsNullOrEmpty(txt_Sueldo.Text) ? -1 : decimal.Parse(txt_Sueldo.Text);
                 empleados1.FechaPago = Convert.ToDateTime(dtp_Pago_Agregar_Empleado.Text);
                 empleados1.FechaContratacion = Convert.ToDateTime(dtp_Ingreso_Agregar_Empleado.Text);
                 empleados1.CodigoJornada = cmb_Jornada_Laboral.SelectedIndex+1;
